llm/requirements.txt
fastapi
uvicorn
transformers
torch
google.generativeai
python-multipart
awsebcli
botocore
boto3
gunicorn
python-dotenv
requests